Cyrillic update
Cyrillic upper and lower cases are now finished for the first five masters. See presentation in PDF format for a complete spacing and cascading text sizes preview.

IMPORTANT: designs of 'bar' or 'slash' characters are  compatible but not yet ready for alternat designs in the bolder master, so they may look clogged in this presentation.

For a full report of the progress read the updated log file.