Converting over to UFR format
I realized the main issue with the way the font was building was that there was not an ExtraLight instance at the 7pt optical size. This was causing issues with fontmake.

To that end, I restored the original Glyphs sources, fixed the ohorn issues (previously fixed by Marc) and now build from there. There are still issues that need resolving (such as implementing the jump feature as rvrn instead).

Still needs review and confirmation that all is working as expected.